refactor: onSubmitComment 로직 useStudylogComment에 넣기
