+++ frontend/src/hooks/Comment/useStudylogComment.ts
@@ -1,3 +1,7 @@
+import { useRef } from 'react';
+import { Editor as ToastEditor } from '@toast-ui/react-editor';
+
+import { ALERT_MESSAGE } from '../../constants';
 import { CommentRequest } from '../../models/Comment';
 import {
   useCreateComment,
@@ -7,6 +11,8 @@ import {
 } from '../queries/comment';
 
 const useStudylogComment = (studylogId: number) => {
+  const editorContentRef = useRef<ToastEditor>(null);
+
   const { data } = useFetchComments(studylogId);
   const comments = data?.data;
 
@@ -26,11 +32,27 @@ const useStudylogComment = (studylogId: number) => {
     deleteCommentMutation.mutate(commentId);
   };
 
+  const onSubmitComment = (event) => {
+    event.preventDefault();
+
+    const content = editorContentRef.current?.getInstance().getMarkdown() || '';
+
+    if (content.length === 0) {
+      alert(ALERT_MESSAGE.NO_CONTENT);
+      return;
+    }
+
+    createComment({ content });
+    editorContentRef.current?.getInstance().setMarkdown('');
+  };
+
   return {
     comments,
+    editorContentRef,
     createComment,
     editComment,
     deleteComment,
+    onSubmitComment,
   };
 };
